Add files via upload
This is my first 'wheels-off' project to, once finished, be a surprise for an incredibly kind dog trainer who gave me the tools, time, and guidance to keep and transform an incredibly challenging dog and help her experience the world and our relationship in far more functional and calm (and continually improving <3) ways.

This initial upload is of just the main index homepage and corresponding CSS file. I am having difficulty figuring out how to create a nice dropdown menu in the navigation bar and haven't, yet, determined how to get the 'dropdown triangle' next to the navbar link with the dropdown portion displaying the same formatting.

Thank you so much for taking a look and any thoughts or tips you have!